Split code into JS modules
